+++ src/crypto/stream/xsalsa20.rs
@@ -1,7 +1,8 @@
 //! `crypto_stream_xsalsa20`, a particular cipher specified in
-//! [Cryptography in NaCl](http://nacl.cr.yp.to/valid.html), Section 7.
+//! [Cryptography in `NaCl`](http://nacl.cr.yp.to/valid.html), Section 7.
 //! This cipher is conjectured to meet the standard notion of
 //! unpredictability.
+
 use ffi::{
     crypto_stream_xsalsa20, crypto_stream_xsalsa20_KEYBYTES, crypto_stream_xsalsa20_NONCEBYTES,
     crypto_stream_xsalsa20_xor, crypto_stream_xsalsa20_xor_ic,
@@ -11,8 +12,8 @@ stream_module!(
     crypto_stream_xsalsa20,
     crypto_stream_xsalsa20_xor,
     crypto_stream_xsalsa20_xor_ic,
-    crypto_stream_xsalsa20_KEYBYTES,
-    crypto_stream_xsalsa20_NONCEBYTES
+    crypto_stream_xsalsa20_KEYBYTES as usize,
+    crypto_stream_xsalsa20_NONCEBYTES as usize
 );
 
 #[cfg(test)]
